Merged Parent theme into Child theme
The way to do it doesn't matter much as in the end it is just one
theme anyway.
The advantage to merge parent in child was, that we will have
less trouble with menu settings and stuff
as that is saved to the current stylesheet handle
which will stay consistent this way.